feat: RAG chat sessions, product media, shop locations, analytics, doc preprocessing

+++ test/integration/modules/shop/shop.service.integration.spec.ts
@@ -2,8 +2,8 @@ import { mockCacheService } from '@/common/utils';
 import { CacheService } from '@/core/cache/cache.service';
 import { ChatEvent, StorefrontView } from '@/modules/analytics/entities';
 import { Order } from '@/modules/order/entities';
-import { Shop } from '@/modules/shop/entities';
-import { ShopRepository } from '@/modules/shop/repositories';
+import { Location, Shop } from '@/modules/shop/entities';
+import { LocationRepository, ShopRepository } from '@/modules/shop/repositories';
 import { ShopService } from '@/modules/shop/shop.service';
 import { User } from '@/modules/user/entities';
 import { getPostgresConnection } from '../../setup';
@@ -34,11 +34,12 @@ describe('ShopService Integration', () => {
           synchronize: true,
           logging: false,
         }),
-        TypeOrmModule.forFeature([Shop, User, ChatEvent, StorefrontView, Order]),
+        TypeOrmModule.forFeature([Shop, Location, User, ChatEvent, StorefrontView, Order]),
       ],
       providers: [
         ShopService,
         ShopRepository,
+        LocationRepository,
         {
           provide: CacheService,
           useValue: mockCacheService(),
@@ -54,13 +55,18 @@ describe('ShopService Integration', () => {
   }, 120000);
 
   afterEach(async () => {
+    if (!dataSource?.isInitialized) {
+      return;
+    }
     await dataSource.query('UPDATE shops SET "ownerId" = NULL');
     await dataSource.query('DELETE FROM users');
     await dataSource.query('DELETE FROM shops');
   });
 
   afterAll(async () => {
-    await app.close();
+    if (app) {
+      await app.close();
+    }
   }, 30000);
 
   describe('create', () => {
